# Replace some thrust usage with cub in torch.unique to gain 2-3x speedup on GPU