--- a/aten/src/ATen/native/cuda/Unique.cu
+++ b/aten/src/ATen/native/cuda/Unique.cu
@@ -6,7 +6,7 @@
 #include <tuple>
 #include <iterator>
 
-#include <cub/device/device_radix_sort.hpp>
+#include <cub/device/device_radix_sort.cuh>
 #include <cub/device/device_run_length_encode.cuh>
 
 #include <THC/THCThrustAllocator.cuh>