--- a/aten/src/ATen/native/cuda/Unique.cu
+++ b/aten/src/ATen/native/cuda/Unique.cu
@@ -5,8 +5,13 @@
 #include <tuple>
 #include <iterator>
 
+#if defined(__HIPCC__)
+#include <cub/device/device_radix_sort.hpp>
+#include <cub/device/device_run_length_encode.hpp>
+#else
 #include <cub/device/device_radix_sort.cuh>
 #include <cub/device/device_run_length_encode.cuh>
+#endif
 
 #include <THC/THCThrustAllocator.cuh>
 #include <thrust/execution_policy.h>