Some of the thrust usages are replaced with cub for `torch.unique` and `torch.unique_consecutive`.

Benchmark code:

```python
import torch
print(torch.__version__)
print()

a = torch.randint(1000, (100000,)).cuda()
torch.cuda.synchronize()

for i in range(100):
    a.unique()

print('single runs:')
torch.cuda.synchronize()
%timeit a.unique(); torch.cuda.synchronize()
%timeit a.unique(return_counts=True); torch.cuda.synchronize()
%timeit a.unique(return_counts=True, return_inverse=True); torch.cuda.synchronize()
%timeit a.unique_consecutive(); torch.cuda.synchronize()
%timeit a.unique_consecutive(return_counts=True); torch.cuda.synchronize()
%timeit a.unique_consecutive(return_counts=True, return_inverse=True); torch.cuda.synchronize()

def run100(f):
    for i in range(100):
        f()

print()
print('100 runs:')
torch.cuda.synchronize()
%timeit run100(lambda: a.unique()); torch.cuda.synchronize()
%timeit run100(lambda: a.unique(return_counts=True)); torch.cuda.synchronize()
%timeit run100(lambda: a.unique(return_counts=True, return_inverse=True)); torch.cuda.synchronize()
%timeit run100(lambda: a.unique_consecutive()); torch.cuda.synchronize()
%timeit run100(lambda: a.unique_consecutive(return_counts=True)); torch.cuda.synchronize()
%timeit run100(lambda: a.unique_consecutive(return_counts=True, return_inverse=True)); torch.cuda.synchronize()
```

Result on master:
```
1.4.0a0+ed20937

single runs:
771 µs ± 60.9 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
879 µs ± 110 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
744 µs ± 17.2 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
126 µs ± 10.6 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
237 µs ± 14.4 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
352 µs ± 71.2 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

100 runs:
71.8 ms ± 3.77 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
95.6 ms ± 17.5 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
96.6 ms ± 17.4 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
10 ms ± 523 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
16.5 ms ± 2.04 ms per loop (mean ± std. dev. of 7 runs, 100 loops each)
34.1 ms ± 5.63 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

With this PR:
```
1.4.0a0+7723340

single runs:
289 µs ± 1.48 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
291 µs ± 1.5 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
393 µs ± 2.41 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
49.5 µs ± 1.14 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
51.2 µs ± 481 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
102 µs ± 539 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

100 runs:
28.1 ms ± 90.1 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
28 ms ± 73.2 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
38.4 ms ± 110 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
4.01 ms ± 45.5 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
4.01 ms ± 41.8 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
9 ms ± 24.7 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```