@kostmo 2019-12-12
## CircleCI build failures summary
As of commit 187f0c1:
* **2/2** failures introduced in this PR



## Detailed failure analysis

One may explore the probable reasons each build failed interactively [on the Dr. CI website](https://dr.pytorch.org/commit-details.html?sha1=187f0c13be3fe29407cd157b0e26126baf19b09e).

### 2 new failures recognized by patterns
The following build failures don't appear to be due to upstream breakage:
#### [![See CircleCI build](https://avatars0.githubusercontent.com/ml/7?s=12)](https://circleci.com/gh/pytorch/pytorch/3912381) pytorch_macos_10_13_py3_test (1/2)
**Step:** "Test" ([details](https://dr.pytorch.org/build-details.html?build_id=72688537))
```
Dec 11 17:00:20   File "test_fake_quant.py", line 7, in <module>
Dec 11 17:00:20     import hypothesis_utils as hu
Dec 11 17:00:20   File "/Users/distiller/project/test/hypothesis_utils.py", line 9, in <module>
Dec 11 17:00:20     from hypothesis.searchstrategy import SearchStrategy
Dec 11 17:00:20 ModuleNotFoundError: No module named 'hypothesis.searchstrategy'
Dec 11 17:00:20 Traceback (most recent call last):
Dec 11 17:00:20   File "test/run_test.py", line 456, in <module>
Dec 11 17:00:20     main()
Dec 11 17:00:20   File "test/run_test.py", line 449, in main
Dec 11 17:00:20     raise RuntimeError(message)
Dec 11 17:00:20 RuntimeError: test_fake_quant failed!
Dec 11 17:00:20 + cleanup
Dec 11 17:00:20 + retcode=1
Dec 11 17:00:20 + set +x
```
#### [![See CircleCI build](https://avatars0.githubusercontent.com/ml/7?s=12)](https://circleci.com/gh/pytorch/pytorch/3912597) pytorch_linux_xenial_py3_clang5_asan_test (2/2)
**Step:** "Test" ([details](https://dr.pytorch.org/build-details.html?build_id=72688702))
```
Dec 12 01:00:34   File "test_nn.py", line 56, in <module>
Dec 12 01:00:34     import hypothesis_utils as hu
Dec 12 01:00:34   File "/var/lib/jenkins/workspace/test/hypothesis_utils.py", line 9, in <module>
Dec 12 01:00:34     from hypothesis.searchstrategy import SearchStrategy
Dec 12 01:00:34 ModuleNotFoundError: No module named 'hypothesis.searchstrategy'
Dec 12 01:00:35 Traceback (most recent call last):
Dec 12 01:00:35   File "test/run_test.py", line 456, in <module>
Dec 12 01:00:35     main()
Dec 12 01:00:35   File "test/run_test.py", line 449, in main
Dec 12 01:00:35     raise RuntimeError(message)
Dec 12 01:00:35 RuntimeError: test_nn failed!
Dec 12 01:00:35 + cleanup
Dec 12 01:00:35 + retcode=1
Dec 12 01:00:35 + set +x
Dec 12 01:00:35 =================== sccache compilation log ===================
Dec 12 01:00:35 =========== If your build fails, please take a look at the log above for possible reasons ===========
Dec 12 01:00:35 Compile requests                  7
Dec 12 01:00:35 Compile requests executed         6
Dec 12 01:00:35 Cache hits                        0
Dec 12 01:00:35 Cache misses                      6
Dec 12 01:00:35 Cache timeouts                    0
```


---
This comment was automatically generated by [Dr. CI](https://github.com/kostmo/circleci-failure-tracker/tree/master/docs/from-pull-request-comment).
Follow [this link to opt-out](https://dr.pytorch.org/admin/comments-opt-out.html) of these comments for your Pull Requests.


Please report bugs/suggestions on the [GitHub issue tracker](https://github.com/kostmo/circleci-failure-tracker/issues).