@zasdfgbnm 2019-12-12
I will refactor the test tomorrow.