@zasdfgbnm 2019-12-12
@iotamudelta Any idea on how to make this work for ROCm?

```
00:33:08 /var/lib/jenkins/workspace/aten/src/ATen/native/hip/Unique.hip:58:7: error: use of undeclared identifier 'cub'
00:33:08       cub::DeviceRadixSort::SortPairs(tmp_storage, tmp_storage_size, self_data, sorted_data_, arange_data, sorted_indices_data, num_inp, 0, sizeof(scalar_t) * 8, stream);
00:33:08       ^
```