@iotamudelta 2019-12-12
This is supported in ROCm, header `cub/device/device_radix_sort.hpp` (relative to hipcub include dir). Could you try including the actual header?

`cub/cub.cuh` is not available on ROCm.