@zasdfgbnm 2019-12-13
@iotamudelta I keep seeing the following error
```
01:54:10 /var/lib/jenkins/workspace/aten/src/ATen/native/hip/Unique.hip:9:10: fatal error: 'cub/device/device_radix_sort.cuh' file not found
01:54:10 #include <cub/device/device_radix_sort.cuh>
01:54:10          ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
01:54:10 1 error generated.
```
And I tried to add
```python
        ("cub/device/device_radix_sort.cuh", ("hipcub/hipcub.hpp", CONV_INCLUDE, API_BLAS)),
        ("cub/device/device_run_length_encode.cuh", ("hipcub/hipcub.hpp", CONV_INCLUDE, API_BLAS)),
```
to `cuda_to_hip_mappings.py`, but now the error becomes
```
02:09:18 /var/lib/jenkins/workspace/aten/src/ATen/native/hip/Unique.hip:60:7: error: use of undeclared identifier 'cub'
02:09:18       cub::DeviceRadixSort::SortPairs(nullptr, tmp_storage_size, self_data, sorted_data_, arange_data, sorted_indices_data, num_inp, 0, sizeof(scalar_t) * 8, stream);
02:09:18       ^
```
Any hint?