@ngimel 2019-12-13
After discussing internally, having 2 paths really does not sound good in terms of code testing and maintenance and compile times (Unique has already been an offender for compilation times). Given that there probably are not cases where unique is a real performance bottleneck (?) and that this PR still does not get rid of all the synchronizations thus does not enable graph capture, I think the drawbacks overweigh the advantages,  and we should abandon this for now. Thank you for you work, @zasdfgbnm, it's really awesome, and testing improvements and bug fixes that you made in the process are really useful too!